HW 2: Complexity of Recursions and Sorting
Problem 5. You are given a list of n tuples (a[i], b[i]), each tuple
represents a segment [ai, bi], i = 1, 2, . . . , n, on a line.
(a)Return the maximum integer k such that some point
on the line is covered with k segments.
For instance, for the following five segments [−0.3, 2.7], [−1, 0.5],
[2.5, 3.3], [−0.7, 2.3], [0.5, 1.7], the answer
is k = 4, because a point 0.5 is covered with 4 segments. Note also
that you do not need to find this point (or points); the only thing
that you have to find is k.
The run-time must be in O(n log n).

Problem 6. Given two integer arrays, return an array
of their intersection. Each element in the result must appear as many
times as it shows in both arrays and you may return the result in any
order.
Example.
Input: array1 = [3,7,2,6,6,9,5,1,1], array2 = [9,4,9,8,6,1,0,6,6]
Output: [6,6,1,9]
[1,6,6,9] or any other permutations are also accepted.
If array do not have common element, the intersection is an empty
list.
The run-time must be in O(n + m), where n and m are sizes of the
arrays.